Sensor test added

+++ src/main/java/com/torresj/apisensorserver/controller/SensorController.java
@@ -1,11 +1,15 @@
 package com.torresj.apisensorserver.controller;
 
+import com.torresj.apisensorserver.exceptions.EntityAlreadyExists;
 import com.torresj.apisensorserver.exceptions.EntityNotFoundException;
 import com.torresj.apisensorserver.models.Sensor;
+import com.torresj.apisensorserver.models.User.Role;
 import com.torresj.apisensorserver.models.Variable;
 import com.torresj.apisensorserver.services.SensorService;
+import com.torresj.apisensorserver.services.UserService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import java.security.Principal;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.data.domain.Page;
@@ -33,8 +37,12 @@ public class SensorController {
   /* Services */
   private SensorService sensorService;
 
-  public SensorController(SensorService sensorService) {
+  private UserService userService;
+
+  public SensorController(SensorService sensorService,
+      UserService userService) {
     this.sensorService = sensorService;
+    this.userService = userService;
   }
 
   @GetMapping
@@ -42,9 +50,15 @@ public class SensorController {
   public ResponseEntity<Page<Sensor>> getSensors(
       @RequestParam(value = "page") int nPage,
       @RequestParam(value = "elements") int elements,
-      @RequestParam(value = "sensorTypeId", required = false) Long sensorTypeId
+      @RequestParam(value = "sensorTypeId", required = false) Long sensorTypeId,
+      Principal principal
   ) {
     try {
+      logger.info("[SENSOR - GET ALL] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info(
           "[SENSOR - GET ALL] Get sensors from DB with page " + nPage + ", elements " + elements
               + ", sensorTypeId " + sensorTypeId);
@@ -53,6 +67,10 @@ public class SensorController {
           : sensorService.getSensors(nPage, elements, sensorTypeId);
 
       return new ResponseEntity<>(page, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR - GET ALL] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (Exception e) {
       logger.error("[SENSOR - GET ALL] Error getting sensors from DB", e);
       throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal error", e);
@@ -61,13 +79,24 @@ public class SensorController {
 
   @GetMapping(value = "/{id}")
   @ApiOperation(value = "Retrieve sensor by id", response = Sensor.class)
-  public ResponseEntity<Sensor> getSensorByID(@PathVariable("id") long id) {
+  public ResponseEntity<Sensor> getSensorByID(@PathVariable("id") long id,
+      Principal principal) {
     try {
+      logger.info("[SENSOR - GET] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)
+          && !sensorService.hasUserVisibilitySensor(principal.getName(), id)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info("[SENSOR - GET] Get sensor from DB with id: " + id);
 
       Sensor sensor = sensorService.getSensor(id);
 
       return new ResponseEntity<>(sensor, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR - GET] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (EntityNotFoundException e) {
       logger.error("[SENSOR - GET] Sensor not found", e);
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sensor not found", e);
@@ -81,8 +110,15 @@ public class SensorController {
   @ApiOperation(value = "Retrieve variables sensor by id", response = Variable.class, responseContainer = "List")
   public ResponseEntity<Page<Variable>> getVariablesSensorByID(@PathVariable("id") long id,
       @RequestParam(value = "page") int nPage,
-      @RequestParam(value = "elements") int elements) {
+      @RequestParam(value = "elements") int elements,
+      Principal principal) {
     try {
+      logger.info("[SENSOR VARIABLES - GET] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)
+          && !sensorService.hasUserVisibilitySensor(principal.getName(), id)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info(
           "[SENSOR VARIABLES - GET] Get variables sensor from DB with id: " + id + " page: " + nPage
               + ", elements: " + elements);
@@ -91,6 +127,10 @@ public class SensorController {
       Page<Variable> page = sensorService.getVariables(id, nPage, elements);
 
       return new ResponseEntity<>(page, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR VARIABLES - GET] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (EntityNotFoundException e) {
       logger.error("[SENSOR VARIABLES - GET] Sensor not found", e);
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sensor not found", e);
@@ -100,11 +140,16 @@ public class SensorController {
     }
   }
 
-  @PutMapping(value = "/{id}/variables")
+  @PutMapping(value = "/{id}/variables/{variableId}")
   @ApiOperation(value = "Add variable to sensor variables list", response = Variable.class, notes = "Variable must exist")
   public ResponseEntity<Variable> addVariablesSensorByID(@PathVariable("id") long id,
-      @RequestBody() long variableId) {
+      @PathVariable("variableId") long variableId, Principal principal) {
     try {
+      logger.info("[SENSOR VARIABLES - ADD] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info(
           "[SENSOR VARIABLES - ADD] Add variable " + variableId + " to variables sensor " + id
               + " list");
@@ -112,20 +157,29 @@ public class SensorController {
       Variable variable = sensorService.addVariable(id, variableId);
 
       return new ResponseEntity<>(variable, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR VARIABLES - GET] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (EntityNotFoundException e) {
-      logger.error("[SENSOR - GET] Sensor or variable not found", e);
+      logger.error("[SENSOR VARIABLES - ADD] Sensor or variable not found", e);
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sensor or variable not found", e);
     } catch (Exception e) {
-      logger.error("[SENSOR - GET] Error getting sensors from DB", e);
+      logger.error("[SENSOR VARIABLES - ADD] Error getting sensors from DB", e);
       throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal error", e);
     }
   }
 
-  @DeleteMapping(value = "/{id}/variables")
+  @DeleteMapping(value = "/{id}/variables/{variableId}")
   @ApiOperation(value = "Delete variable from sensor variables list", response = Variable.class, notes = "Variable must exist")
   public ResponseEntity<Variable> deleteVariablesSensorByID(@PathVariable("id") long id,
-      @RequestBody() long variableId) {
+      @PathVariable("variableId") long variableId, Principal principal) {
     try {
+      logger.info("[SENSOR VARIABLES - REMOVE] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info(
           "[SENSOR VARIABLES - REMOVE] Remove variable " + variableId + " from variables sensor "
               + id
@@ -134,6 +188,10 @@ public class SensorController {
       Variable variable = sensorService.removeVariable(id, variableId);
 
       return new ResponseEntity<>(variable, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR VARIABLES - REMOVE] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (EntityNotFoundException e) {
       logger.error("[SENSOR VARIABLES - REMOVE] Sensor or variable not found", e);
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sensor or variable not found", e);
@@ -145,12 +203,21 @@ public class SensorController {
 
   @PutMapping
   @ApiOperation(value = "Update sensor", response = Sensor.class, notes = "SensorType and House must exist. House can be null")
-  public ResponseEntity<Sensor> update(@RequestBody() Sensor sensor) {
+  public ResponseEntity<Sensor> update(@RequestBody() Sensor sensor, Principal principal) {
     try {
+      logger.info("[SENSOR - UPDATE] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info("[SENSOR - UPDDATE] Updating sensor -> " + sensor);
       Sensor sensorRegister = sensorService.update(sensor);
 
       return new ResponseEntity<>(sensorRegister, HttpStatus.CREATED);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR - UPDATE] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (Exception e) {
       logger.error("[SENSOR - UPDATE] Error updating sensor", e);
       throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal error", e);
@@ -159,12 +226,23 @@ public class SensorController {
 
   @PostMapping
   @ApiOperation(value = "Register sensor", response = Sensor.class)
-  public ResponseEntity<Sensor> register(@RequestBody() Sensor sensor) {
+  public ResponseEntity<Sensor> register(@RequestBody() Sensor sensor, Principal principal) {
     try {
+      logger.info("[SENSOR - REGISTER] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN, Role.STATION)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info("[SENSOR - REGISTER] Registering sensor -> " + sensor);
       Sensor sensorRegister = sensorService.register(sensor);
-
       return new ResponseEntity<>(sensorRegister, HttpStatus.CREATED);
+    } catch (EntityAlreadyExists e) {
+      logger.info("[SENSOR - REGISTER] Entity already exists and it hasn't been modified/created");
+      return new ResponseEntity<>((Sensor) e.getObject(), HttpStatus.ACCEPTED);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR - REGISTER] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (Exception e) {
       logger.error("[SENSOR - REGISTER] Error registering sensor", e);
       throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Internal error", e);
@@ -173,13 +251,22 @@ public class SensorController {
 
   @DeleteMapping(value = "/{id}")
   @ApiOperation(value = "Delete sensor", response = Sensor.class)
-  public ResponseEntity<Sensor> delete(@PathVariable("id") long id) {
+  public ResponseEntity<Sensor> delete(@PathVariable("id") long id, Principal principal) {
     try {
+      logger.info("[SENSOR - REMOVE] Check user permission");
+      if (!userService.isUserAllowed(principal.getName(), Role.ADMIN)) {
+        throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+            "user Not have permission for this endpoint");
+      }
       logger.info("[SENSOR - REMOVE] Remove sensor from DB with id: " + id);
 
       Sensor sensor = sensorService.removeSensor(id);
 
       return new ResponseEntity<>(sensor, HttpStatus.OK);
+    } catch (ResponseStatusException e) {
+      logger.error("[SENSOR - REMOVE] user Not have permission for this endpoint");
+      throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+          e.getReason(), e);
     } catch (EntityNotFoundException e) {
       logger.error("[SENSOR - REMOVE] Sensor not found", e);
       throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Sensor not found", e);
